+++ create_htaccess.sh
@@ -6,8 +6,10 @@ prod_name=$3
 
 echo -e "*** HTACCESS ***"
 
-echo -e "\t--- supression du .htaccess (si existe) ---"
-rm $path/.htaccess
+if [ -f "$path/.htaccess" ];then
+  echo -e "\t--- supression du .htaccess ---"
+  rm $path/.htaccess
+fi
 
 echo -e "\t--- création du .htaccess ---"
